refactor: 로또 생성 역할Machine으로 이동
